Implement itoa and friends
This implements ultoa, utoa, ltoa, itoa and dtostrf. These are normally
defined by stdlib.h in avr-libc, but not by libc on x86 or x86_64.

The declarations in xtoa.h are taken from avr-libc, but avr-libc only
contains an assembly implementation for these functions, so the
functions are implemented from scratch.